Restructure util to change underscore usage
Basically instead of underscores for each section under util, the whole
util section is underscored.
