+++ weasel/_util/compat.py
