+++ weasel/_util/errors.py
