+++ weasel/_util/schema.py
